AddNewTransaction

+++ AILendTreasury.Data/Entities/Interfaces/ITransaction.cs
@@ -8,8 +8,8 @@ namespace AILendTreasury.Data.Entities.Interfaces
     {
         string SoldCurrency { get; set; }
         string BoughtCurrency { get; set; }
-        long SoldAmount { get; set; }
-        long BoughtAmount { get; set; }
+        double SoldAmount { get; set; }
+        double BoughtAmount { get; set; }
         float ExchangeRate { get; set; }
         DateTime CreatedDate { get; set; }
     }
