+++ AILendTreasury.Data/Entities/Transaction.cs
@@ -9,8 +9,8 @@ namespace AILendTreasury.Data.Entities
     {
         public string SoldCurrency { get; set; }
         public string BoughtCurrency { get; set; }
-        public long SoldAmount { get; set; }
-        public long BoughtAmount { get; set; }
+        public double SoldAmount { get; set; }
+        public double BoughtAmount { get; set; }
         public float ExchangeRate { get; set; }
         public DateTime CreatedDate { get; set; }
     }
